Moved duplicate put requests to a utils folder
