+++ src/GuestPages/EditRSVP.js
@@ -6,6 +6,7 @@ import axios from 'axios';
 import Success from './Success';
 import PlumFilledButton from '../Common/styled/buttonstyles/PlumFilledButton';
 import PlumButton from '../Common/styled/buttonstyles/PlumButton';
+import updateRSVP from './utils/updateRSVP';
 
 const { URL } = require('../../config/private.config');
 
@@ -27,22 +28,7 @@ const EditRsvp = ({ updateStep, rsvpData }) => {
     if (isAttending === null) {
       alert('Please select one of the options for attending');
     } else {
-      const data = {
-        rsvpStatus: isAttending,
-        rsvpNote: message,
-      };
-
-      axios({
-        method: 'put',
-        url: `${URL}/events/61b79b9e0ac02dbe3e12fd1b/guests/${rsvpData._id}`,
-        data,
-      })
-        .then(() => {
-          updateResponse(true);
-        })
-        .catch((err) => {
-          console.log(err);
-        });
+      updateRSVP(isAttending, message, rsvpData._id, updateResponse);
     }
   };
